add custom comparator with documentation, reuse that in both planning and validation
