+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/table/action/savepoint/SavepointHelpers.java
@@ -18,23 +18,16 @@
 
 package org.apache.hudi.table.action.savepoint;
 
-import org.apache.hudi.common.model.HoodieTableType;
-import org.apache.hudi.common.table.HoodieTableVersion;
 import org.apache.hudi.common.table.timeline.HoodieInstant;
 import org.apache.hudi.common.table.timeline.HoodieTimeline;
-import org.apache.hudi.common.table.timeline.InstantComparator;
 import org.apache.hudi.common.util.Option;
 import org.apache.hudi.common.util.ValidationUtils;
 import org.apache.hudi.exception.HoodieRollbackException;
 import org.apache.hudi.table.HoodieTable;
+import org.apache.hudi.table.action.rollback.RestoreInstantComparator;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.util.Comparator;
-import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 public class SavepointHelpers {
 
@@ -62,8 +55,7 @@ public class SavepointHelpers {
         .getWriteTimeline()
         .filterCompletedAndCompactionInstants()
         .getInstantsAsStream()
-        .max(new SavepointInstantComparator(table.getMetaClient().getTableConfig().getTableVersion().greaterThanOrEquals(HoodieTableVersion.EIGHT),
-            table.getMetaClient().getTimelineLayout().getInstantComparator(), table.getMetaClient().getTableType())));
+        .max(new RestoreInstantComparator(table.getMetaClient())));
     ValidationUtils.checkArgument(lastInstant.isPresent());
     ValidationUtils.checkArgument(lastInstant.get().requestedTime().equals(savepointTime),
         () -> savepointTime + " is not the last commit after restoring to savepoint, last commit was "
@@ -75,33 +67,6 @@ public class SavepointHelpers {
     boolean isSavepointPresent = table.getCompletedSavepointTimeline().containsInstant(savePoint);
     if (!isSavepointPresent) {
       throw new HoodieRollbackException("No savepoint for instantTime " + savepointTime);
-    }
-  }
-
-  private static class SavepointInstantComparator implements Comparator<HoodieInstant> {
-    private static final Set<String> COMPACTION_ACTIONS = Stream.of(HoodieTimeline.COMPACTION_ACTION, HoodieTimeline.COMMIT_ACTION).collect(Collectors.toSet());
-    private final boolean tableVersion8OrLater;
-    private final InstantComparator instantComparator;
-    private final HoodieTableType tableType;
-
-    public SavepointInstantComparator(boolean tableVersion8OrLater, InstantComparator instantComparator, HoodieTableType tableType) {
-      this.tableVersion8OrLater = tableVersion8OrLater;
-      this.instantComparator = instantComparator;
-      this.tableType = tableType;
-    }
-
-    @Override
-    public int compare(HoodieInstant o1, HoodieInstant o2) {
-      if (tableVersion8OrLater || tableType == HoodieTableType.COPY_ON_WRITE) {
-        return instantComparator.completionTimeOrderedComparator().compare(o1, o2);
-      } else {
-        // Do to special handling of compaction instants, we need to use requested time based comparator for compaction instants but completion time based comparator for others
-        if (COMPACTION_ACTIONS.contains(o1.getAction()) || COMPACTION_ACTIONS.contains(o2.getAction())) {
-          return instantComparator.requestedTimeOrderedComparator().compare(o1, o2);
-        } else {
-          return instantComparator.completionTimeOrderedComparator().compare(o1, o2);
-        }
-      }
     }
   }
 }
